Sample app to demonstrate iOS8 specific animation bug (regression) in - (void)dismissViewControllerAnimated:(BOOL)flag             completion:(void (^)(void))completion
According to Apple documentation:
If you present several view controllers in succession,
thus building a stack of presented view controllers,
calling this method on a view controller lower in the stack
dismisses its immediate child view controller and
all view controllers above that child on the stack.
When this happens, only the top-most view is dismissed in an animated fashion;
any intermediate view controllers are simply removed from the stack.

However, on iOS8 top-most view is dimission without animation and view that was
presented before top-most is animated.
iOS7 doesn not have this bug, so it must imply a regression in this code.