Demonstrate that iOS8 navigation bar won't change heigh in compact mode when presented from view hierarchy that does not belong to the original window.
With iOS8 navigation bar will changes its height to 32 pixels from 44 pixels.

This is described in design guidelines as:
"iOS automatically makes various layout changes when the size classes
of a display environment change. For example, when the vertical size
class changes from compact to regular, navigation bars and toolbars
automatically become taller."

However, the same height adjustment doesn't apply to navigation bars
that are added to view controllers outside of the first original window key.

It makes it harder to adjust