Demonstration of the fact that UIView of the UIBarButtonItem presented in UIToolbar is recreated in some cases upon rotation on iPhone, but not on iPad.
This can't be considered Apple bug, because Apple does not expose
direct interface to UIView of UIBarButtonItem. The way I get it here
is kind of hacky.

However, it is still odd that iPad and iPhone behave differently
for UIBarButtonItems in UIToolbar upon rotation.
There is no such behavior for UIBarButton for UINavigationBar.